Update tests structure. Each app module has own tests folder
